Set lock_timeout on session when disable_ddl_transaction
`SET LOCAL lock_timeout` requires a transaction to be present which
means we can't intervene when `disable_ddl_transaction!` has been
called.

This adds support for this case, using `SET lock_timeout`. Since
this will impact the whole session, care is taken to ensure the
setting is returned to its original value.